requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-Login==0.5.0
mysqlclient==2.0.3
python-dotenv==0.19.0
pytest==6.2.5
WTForms==2.3.3
Flask-Migrate==3.1.0
email-validator==1.1.3 